Restructure auth
