+++ auth/util.hash.go
